requirements.txt
google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
chromadb
python-dateutil
dateparser
python-dotenv
